Split Mac resource directories into "normal" and "-frozen"
Previously, the folder src/main/resources/mac was meant to contain a
subfolder structure that mimics the layout of the frozen .app folder.
For example: Contents/Info.plist. If you wanted to add a "normal"
resource file, you had to place it in .../mac/Contents/Resources - but
then fbs's get_resource(...) function was unable to find it (because it
expected the file to be at just .../mac, as for the other OSs).

Now, there are "normal" resources and those that are only required for
the frozen version of an app. The former go into src/main/resources/mac
.../windows etc., whereas the latter go into .../mac-frozen et al.